Working adding metor ffrrom buildings

+++ game/managers/board_sub_managers/BuildingsEntityManager.py
@@ -11,6 +11,7 @@ from ...dataclasses.FieldDataClass import Fieldv2
 from ...gui.common.Config import Config
 from ...utilities.utilities import calc_distance
 from ..AbstractManager import AbstractManager
+import logging
 
 
 class BuildingsEntityManager(AbstractManager):
@@ -29,14 +30,12 @@ class BuildingsEntityManager(AbstractManager):
         self._coins = DataCache.get_value("coins")
         self.fields_status: dict[str, Fieldv2] = DataCache.get_value("fields_status")
         self.building_location = DataCache.get_value("new_building")
-        self.buildings_status = DataCache.get_value("buildings_status")
         return super().read_cache_values()
 
     def save_cache_values(self):
         DataCache.set_value("coins", self._coins)
         DataCache.set_value("fields_status", self.fields_status)
         DataCache.set_value("new_building", self.building_location)
-        DataCache.set_value("buildings_status", self.buildings_status)
         return super().save_cache_values()
 
     @property
@@ -60,24 +59,20 @@ class BuildingsEntityManager(AbstractManager):
             self.new_metro_decider()
             return
         for field_id, field_data in self.fields_status.items():
-            if field_data.owner == self._act_player:
+            if field_data.type == "island" and field_data.owner == self._act_player:
                 for i in range(len(self.field_config[field_id]["small"])):
-                    if field_data.small_building and not field_data.small_building[str(i + 1)]:
-                        temp_loc = calc_distance(self.building_location, self.field_config[island_id]["small"][i])
+                    if field_data.buildings and not field_data.buildings[str(i + 1)]:
+                        temp_loc = calc_distance(self.building_location, self.field_config[field_id]["small"][i])
                         new_loc = [field_id, str(i + 1)] if temp_loc < closest_distance else new_loc
                         closest_distance = temp_loc if temp_loc < closest_distance else closest_distance
 
         if closest_distance < 50 and self._coins[self._act_player] >= 2:
             self._coins[self._act_player] -= 2
-            temp_id = self.generate_unique_id()
-            self.buildings_status[temp_id] = Building(
+            self.fields_status[new_loc[0]].buildings[new_loc[1]] = Building(  # type: ignore
+                self.generate_unique_id(),
                 self._act_hero,
-                self.field_config[new_loc[0]]["small"][int(new_loc[1]) - 1],
-                new_loc[0],
-                new_loc[1]
+                self.field_config[new_loc[0]]["small"][int(new_loc[1]) - 1]
             )
-            if self.fields_status[new_loc[0]].small_building is not None:
-                self.fields_status[new_loc[0]].small_building[new_loc[1]] = self._act_hero  # type: ignore
         else:
             DataCache.set_value("reset_building", True)
         self.save_cache_values()
@@ -85,23 +80,28 @@ class BuildingsEntityManager(AbstractManager):
     def new_metro_decider(self):
         new_loc = ""
         closest_distance = Config.boards.default_max_len
-
-        for island_id, island_data in self._islands_status.items():
-            if island_data.owner == self._act_player:
-                if not island_data.metropolis:
-                    temp_loc = calc_distance(self.building_location, self.field_config[island_id]["big"])
-                    new_loc = island_id if temp_loc < closest_distance else new_loc
-                    closest_distance = temp_loc if temp_loc < closest_distance else closest_distance
-
-        if closest_distance < 50:
-            temp_id = self.generate_unique_id()
-            self.delete_buildings()
-            self.buildings_status[temp_id] = Building(
-                "metro",
-                self.field_config[new_loc]["big"],
-                new_loc
+        for field_id, field_data in self.fields_status.items():
+            if field_data.owner != self._act_player or field_data.metropolis[0] or field_data.type != "island":
+                continue
+            temp_loc = calc_distance(self.building_location, self.field_config[field_id]["big"])
+            new_loc = field_id if temp_loc < closest_distance else new_loc
+            closest_distance = temp_loc if temp_loc < closest_distance else closest_distance
+        if (
+            closest_distance < 50
+            and self.check_if_build_metro_possible(
+                new_loc,
+                list(DataCache.get_value("building_to_delete").values())
             )
-            self._islands_status[new_loc].metropolis = True
+        ):
+            self.delete_buildings()
+            self.fields_status[new_loc].metropolis = (
+                True,
+                Building(
+                    self.generate_unique_id(),
+                    "metro",
+                    self.field_config[new_loc]["big"]
+                )
+            )
         else:
             DataCache.set_value("reset_building", True)
         self.save_cache_values()
@@ -125,12 +125,27 @@ class BuildingsEntityManager(AbstractManager):
 
     def delete_buildings(self):
         buildings_to_delete: dict[str, int] = DataCache.get_value("building_to_delete")
-        DataCache.set_value("entity_delete", list(buildings_to_delete.values()))
+        entity_to_delete = DataCache.get_value("entity_delete")
+        DataCache.set_value("entity_delete", list(buildings_to_delete.values()) + entity_to_delete)
+
         delete_buildings_id = list(buildings_to_delete.values())
-        for building_id, building_data in self.buildings_status.items():
-            if building_id in delete_buildings_id:
-                self._islands_status[building_data.island].small_building[building_data.place] = ""  # type: ignore
-        for building_id in buildings_to_delete.values():
-            self.buildings_status.pop(building_id)
+        for field_id, field_data in self.fields_status.items():
+            if isinstance(field_data.buildings, dict):
+                for place, building in field_data.buildings.items():
+                    if building and building._id in delete_buildings_id:
+                        field_data.buildings[place] = None
         DataCache.set_value("building_to_delete", {})
         DataCache.set_value("metro_building", False)
+
+    def check_if_build_metro_possible(self, field_id, building_to_delete_ids) -> bool:
+        islands_config = Config.boards.islands_config[str(DataCache.get_value("num_of_players"))]
+        for place in islands_config[field_id]["buildings"]["big"][0]:
+            if not self.fields_status[field_id].buildings[place] or self.fields_status[field_id].buildings[place]._id in building_to_delete_ids:  # type: ignore
+                continue
+            else:
+                logging.info("Deleting building on metro place!")
+                entity_to_delete: list = DataCache.get_value("entity_delete")
+                entity_to_delete.append(self.fields_status[field_id].buildings[place]._id)  # type: ignore
+                self.fields_status[field_id].buildings[place] = None  # type: ignore
+                continue
+        return True
